[SYCL][Graph] Refactor unit tests into multiple files
- Split unit tests into multiple files for easier maintenance
