+++ sycl/unittests/Extensions/CMakeLists.txt
@@ -7,7 +7,7 @@ add_sycl_unittest(ExtensionsTests OBJECT
   USMMemcpy2D.cpp
   DeviceGlobal.cpp
   OneAPISubGroupMask.cpp
-  CommandGraph.cpp
   USMP2P.cpp
 )
 
+add_subdirectory(CommandGraph)
