+++ sycl/unittests/Extensions/CommandGraph/CMakeLists.txt
@@ -1,0 +1,11 @@
+set(CMAKE_CXX_EXTENSIONS OFF)
+
+add_sycl_unittest(CommandGraphExtensionTests OBJECT
+  Barrier.cpp
+  CommandGraph.cpp
+  Exceptions.cpp
+  InOrderQueue.cpp
+  MultiThreaded.cpp
+  Queries.cpp
+  Subgraph.cpp
+)
